pyproject.toml
[project]
name = "math-assistant-prototype"
version = "0.1.0"
description = "DSPy + DeepAgents CLI integration prototype with simple math assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy-ai>=2.4.0",
    "deepagents>=0.1.0",
    "langchain-openai>=0.1.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
math-cli = "math_cli:main"
optimize-prompts = "optimizer:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"